Completed some To-Dos

+++ AntiTask.java
@@ -1,6 +1,6 @@
 public class AntiTask extends Task{
     
-    AntiTask(String name, String type, float startTime, float duration, int date){
+    AntiTask(String name, String type, double startTime, double duration, int date){
         super(name, type, startTime, duration, date);
     }
 }
